+++ src/Divider.jsx
@@ -1,26 +1,19 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import createReactClass from 'create-react-class';
 
-const Divider = createReactClass({
-  displayName: 'Divider',
-
-  propTypes: {
-    disabled: PropTypes.bool,
+export default class Divider extends React.Component {
+  static propTypes = {
     className: PropTypes.string,
     rootPrefixCls: PropTypes.string,
-  },
+  };
 
-  getDefaultProps() {
-    return {
-      disabled: true,
-    };
-  },
+  static defaultProps = {
+    // To fix keyboard UX.
+    disabled: true,
+  };
 
   render() {
     const { className = '', rootPrefixCls } = this.props;
     return <li className={`${className} ${rootPrefixCls}-item-divider`}/>;
-  },
-});
-
-export default Divider;
+  }
+}