+++ src/SubMenu.jsx
@@ -1,19 +1,42 @@
 import React from 'react';
+import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types';
-import createReactClass from 'create-react-class';
-import SubPopupMenu from './SubPopupMenu';
+import Trigger from 'rc-trigger';
 import KeyCode from 'rc-util/lib/KeyCode';
-import classnames from 'classnames';
-import { noop, loopMenuItemRecusively } from './util';
+import classNames from 'classnames';
+import { connect } from 'mini-store';
+import SubPopupMenu from './SubPopupMenu';
+import placements from './placements';
+import Animate from 'rc-animate';
+import {
+  noop,
+  loopMenuItemRecursively,
+  getMenuIdFromSubMenuEventKey,
+  menuAllProps,
+} from './util';
 
 let guid = 0;
 
-/* eslint react/no-is-mounted:0 */
+const popupPlacementMap = {
+  horizontal: 'bottomLeft',
+  vertical: 'rightTop',
+  'vertical-left': 'rightTop',
+  'vertical-right': 'leftTop',
+};
 
-const SubMenu = createReactClass({
-  displayName: 'SubMenu',
+const updateDefaultActiveFirst = (store, eventKey, defaultActiveFirst) => {
+  const menuId = getMenuIdFromSubMenuEventKey(eventKey);
+  const state = store.getState();
+  store.setState({
+    defaultActiveFirst: {
+      ...state.defaultActiveFirst,
+      [menuId]: defaultActiveFirst,
+    },
+  });
+};
 
-  propTypes: {
+export class SubMenu extends React.Component {
+  static propTypes = {
     parentMenu: PropTypes.object,
     title: PropTypes.node,
     children: PropTypes.any,
@@ -24,64 +47,103 @@ const SubMenu = createReactClass({
     rootPrefixCls: PropTypes.string,
     eventKey: PropTypes.string,
     multiple: PropTypes.bool,
-    active: PropTypes.bool,
+    active: PropTypes.bool, // TODO: remove
+    onItemHover: PropTypes.func,
     onSelect: PropTypes.func,
-    closeSubMenuOnMouseLeave: PropTypes.bool,
-    openSubMenuOnMouseEnter: PropTypes.bool,
+    triggerSubMenuAction: PropTypes.string,
     onDeselect: PropTypes.func,
     onDestroy: PropTypes.func,
-    onItemHover: PropTypes.func,
     onMouseEnter: PropTypes.func,
     onMouseLeave: PropTypes.func,
     onTitleMouseEnter: PropTypes.func,
     onTitleMouseLeave: PropTypes.func,
     onTitleClick: PropTypes.func,
-  },
+    popupOffset: PropTypes.array,
+    isOpen: PropTypes.bool,
+    store: PropTypes.object,
+    mode: PropTypes.oneOf(['horizontal', 'vertical', 'vertical-left', 'vertical-right', 'inline']),
+    manualRef: PropTypes.func,
+  };
 
-  mixins: [require('./SubMenuStateMixin')],
+  static defaultProps = {
+    onMouseEnter: noop,
+    onMouseLeave: noop,
+    onTitleMouseEnter: noop,
+    onTitleMouseLeave: noop,
+    onTitleClick: noop,
+    manualRef: noop,
+    mode: 'vertical',
+    title: '',
+  };
 
-  getDefaultProps() {
-    return {
-      onMouseEnter: noop,
-      onMouseLeave: noop,
-      onTitleMouseEnter: noop,
-      onTitleMouseLeave: noop,
-      onTitleClick: noop,
-      title: '',
-    };
-  },
+  constructor(props) {
+    super(props);
+    const store = props.store;
+    const eventKey = props.eventKey;
+    const defaultActiveFirst = store.getState().defaultActiveFirst;
 
-  getInitialState() {
-    this.isSubMenu = 1;
-    return {
-      defaultActiveFirst: false,
-    };
-  },
+    this.isRootMenu = false;
+
+    let value = false;
+
+    if (defaultActiveFirst) {
+      value = defaultActiveFirst[eventKey];
+    }
+
+    updateDefaultActiveFirst(store, eventKey, value);
+  }
+
+  componentDidMount() {
+    this.componentDidUpdate();
+  }
+
+  componentDidUpdate() {
+    const { mode, parentMenu, manualRef } = this.props;
+
+    // invoke customized ref to expose component to mixin
+    if (manualRef) {
+      manualRef(this);
+    }
+
+    if (mode !== 'horizontal' || !parentMenu.isRootMenu || !this.props.isOpen) {
+      return;
+    }
+
+    this.minWidthTimeout = setTimeout(() => this.adjustWidth(), 0);
+  }
 
   componentWillUnmount() {
-    const { onDestroy, eventKey, parentMenu } = this.props;
+    const { onDestroy, eventKey } = this.props;
     if (onDestroy) {
       onDestroy(eventKey);
     }
-    if (parentMenu.subMenuInstance === this) {
-      this.clearSubMenuTimers();
+
+    /* istanbul ignore if */
+    if (this.minWidthTimeout) {
+      clearTimeout(this.minWidthTimeout);
     }
-  },
 
-  onDestroy(key) {
+    /* istanbul ignore if */
+    if (this.mouseenterTimeout) {
+      clearTimeout(this.mouseenterTimeout);
+    }
+  }
+
+  onDestroy = (key) => {
     this.props.onDestroy(key);
-  },
+  };
 
-  onKeyDown(e) {
+  onKeyDown = (e) => {
     const keyCode = e.keyCode;
     const menu = this.menuInstance;
-    const isOpen = this.isOpen();
+    const {
+      isOpen,
+      store,
+    } = this.props;
 
     if (keyCode === KeyCode.ENTER) {
       this.onTitleClick(e);
-      this.setState({
-        defaultActiveFirst: true,
-      });
+      updateDefaultActiveFirst(store, this.props.eventKey, true);
       return true;
     }
 
@@ -90,9 +152,8 @@ const SubMenu = createReactClass({
         menu.onKeyDown(e);
       } else {
         this.triggerOpenChange(true);
-        this.setState({
-          defaultActiveFirst: true,
-        });
+        // need to update current menu's defaultActiveFirst value
+        updateDefaultActiveFirst(store, this.props.eventKey, true);
       }
       return true;
     }
@@ -113,228 +174,173 @@ const SubMenu = createReactClass({
     if (isOpen && (keyCode === KeyCode.UP || keyCode === KeyCode.DOWN)) {
       return menu.onKeyDown(e);
     }
-  },
+  };
 
-  onOpenChange(e) {
+  onOpenChange = (e) => {
     this.props.onOpenChange(e);
-  },
+  };
 
-  onMouseEnter(e) {
-    const props = this.props;
-    this.clearSubMenuLeaveTimer();
-    props.onMouseEnter({
-      key: props.eventKey,
+  onPopupVisibleChange = (visible) => {
+    this.triggerOpenChange(visible, visible ? 'mouseenter' : 'mouseleave');
+  };
+
+  onMouseEnter = (e) => {
+    const { eventKey: key, onMouseEnter, store } = this.props;
+    updateDefaultActiveFirst(store, this.props.eventKey, false);
+    onMouseEnter({
+      key,
       domEvent: e,
     });
-  },
+  };
 
-  onMouseLeave(e) {
-    const { props } = this;
-    const { parentMenu, eventKey } = props;
+  onMouseLeave = (e) => {
+    const {
+      parentMenu,
+      eventKey,
+      onMouseLeave,
+    } = this.props;
     parentMenu.subMenuInstance = this;
-    parentMenu.subMenuLeaveFn = () => {
-      // leave whole sub tree
-      // still active
-      if (props.mode !== 'inline') {
-        const isOpen = this.isOpen();
-        if (isOpen && props.closeSubMenuOnMouseLeave && props.active) {
-          props.onItemHover({
-            key: eventKey,
-            item: this,
-            hover: false,
-            trigger: 'mouseleave',
-            openChanges: [{
-              key: eventKey,
-              item: this,
-              trigger: 'mouseleave',
-              open: false,
-            }],
-          });
-        } else {
-          if (props.active) {
-            props.onItemHover({
-              key: eventKey,
-              item: this,
-              hover: false,
-              trigger: 'mouseleave',
-            });
-          }
-          if (isOpen && props.closeSubMenuOnMouseLeave) {
-            this.triggerOpenChange(false);
-          }
-        }
-      }
-      // trigger mouseleave
-      props.onMouseLeave({
-        key: eventKey,
-        domEvent: e,
-      });
-    };
-    // prevent popup menu and submenu gap
-    parentMenu.subMenuLeaveTimer = setTimeout(parentMenu.subMenuLeaveFn, 100);
-  },
+    onMouseLeave({
+      key: eventKey,
+      domEvent: e,
+    });
+  };
 
-  onTitleMouseEnter(domEvent) {
-    const props = this.props;
-    const { parentMenu, eventKey: key } = props;
-    const item = this;
-    this.clearSubMenuTitleLeaveTimer();
-    if (parentMenu.menuItemInstance) {
-      parentMenu.menuItemInstance.clearMenuItemMouseLeaveTimer();
-    }
-    const openChanges = [];
-    if (props.openSubMenuOnMouseEnter) {
-      openChanges.push({
-        key,
-        item,
-        trigger: 'mouseenter',
-        open: true,
-      });
-    }
-    props.onItemHover({
+  onTitleMouseEnter = (domEvent) => {
+    const { eventKey: key, onItemHover, onTitleMouseEnter } = this.props;
+    onItemHover({
       key,
-      item,
       hover: true,
-      trigger: 'mouseenter',
-      openChanges,
     });
-    this.setState({
-      defaultActiveFirst: false,
-    });
-    props.onTitleMouseEnter({
+    onTitleMouseEnter({
       key,
       domEvent,
     });
-  },
+  };
 
-  onTitleMouseLeave(e) {
-    const { props } = this;
-    const { parentMenu, eventKey } = props;
+  onTitleMouseLeave = (e) => {
+    const { parentMenu, eventKey, onItemHover, onTitleMouseLeave } = this.props;
     parentMenu.subMenuInstance = this;
-    parentMenu.subMenuTitleLeaveFn = () => {
-      // leave whole sub tree
-      // still active
-      if (props.mode === 'inline' && props.active) {
-        props.onItemHover({
-          key: eventKey,
-          item: this,
-          hover: false,
-          trigger: 'mouseleave',
-        });
-      }
-      props.onTitleMouseLeave({
-        key: props.eventKey,
-        domEvent: e,
-      });
-    };
-    parentMenu.subMenuTitleLeaveTimer = setTimeout(parentMenu.subMenuTitleLeaveFn, 100);
-  },
+    onItemHover({
+      key: eventKey,
+      hover: false,
+    });
+    onTitleMouseLeave({
+      key: eventKey,
+      domEvent: e,
+    });
+  };
 
-  onTitleClick(e) {
+  onTitleClick = (e) => {
     const { props } = this;
     props.onTitleClick({
       key: props.eventKey,
       domEvent: e,
     });
-    if (props.openSubMenuOnMouseEnter) {
+    if (props.triggerSubMenuAction === 'hover') {
       return;
     }
-    this.triggerOpenChange(!this.isOpen(), 'click');
-    this.setState({
-      defaultActiveFirst: false,
-    });
-  },
+    this.triggerOpenChange(!props.isOpen, 'click');
+    updateDefaultActiveFirst(props.store, this.props.eventKey, false);
+  };
 
-  onSubMenuClick(info) {
+  onSubMenuClick = (info) => {
     this.props.onClick(this.addKeyPath(info));
-  },
+  };
 
-  onSelect(info) {
+  onSelect = (info) => {
     this.props.onSelect(info);
-  },
+  };
 
-  onDeselect(info) {
+  onDeselect = (info) => {
     this.props.onDeselect(info);
-  },
+  };
 
-  getPrefixCls() {
+  getPrefixCls = () => {
     return `${this.props.rootPrefixCls}-submenu`;
-  },
+  };
 
-  getActiveClassName() {
+  getActiveClassName = () => {
     return `${this.getPrefixCls()}-active`;
-  },
+  };
 
-  getDisabledClassName() {
+  getDisabledClassName = () => {
     return `${this.getPrefixCls()}-disabled`;
-  },
+  };
 
-  getSelectedClassName() {
+  getSelectedClassName = () => {
     return `${this.getPrefixCls()}-selected`;
-  },
+  };
 
-  getOpenClassName() {
+  getOpenClassName = () => {
     return `${this.props.rootPrefixCls}-submenu-open`;
-  },
+  };
 
-  saveMenuInstance(c) {
+  saveMenuInstance = (c) => {
+    // children menu instance
     this.menuInstance = c;
-  },
+  };
 
-  addKeyPath(info) {
+  addKeyPath = (info) => {
     return {
       ...info,
       keyPath: (info.keyPath || []).concat(this.props.eventKey),
     };
-  },
+  };
 
-  triggerOpenChange(open, type) {
+  triggerOpenChange = (open, type) => {
     const key = this.props.eventKey;
-    this.onOpenChange({
-      key,
-      item: this,
-      trigger: type,
-      open,
-    });
-  },
-
-  clearSubMenuTimers() {
-    this.clearSubMenuLeaveTimer();
-    this.clearSubMenuTitleLeaveTimer();
-  },
-
-  clearSubMenuTitleLeaveTimer() {
-    const parentMenu = this.props.parentMenu;
-    if (parentMenu.subMenuTitleLeaveTimer) {
-      clearTimeout(parentMenu.subMenuTitleLeaveTimer);
-      parentMenu.subMenuTitleLeaveTimer = null;
-      parentMenu.subMenuTitleLeaveFn = null;
+    const openChange = () => {
+      this.onOpenChange({
+        key,
+        item: this,
+        trigger: type,
+        open,
+      });
+    };
+    if (type === 'mouseenter') {
+      // make sure mouseenter happen after other menu item's mouseleave
+      this.mouseenterTimeout = setTimeout(() => {
+        openChange();
+      }, 0);
+    } else {
+      openChange();
     }
-  },
+  }
 
-  clearSubMenuLeaveTimer() {
-    const parentMenu = this.props.parentMenu;
-    if (parentMenu.subMenuLeaveTimer) {
-      clearTimeout(parentMenu.subMenuLeaveTimer);
-      parentMenu.subMenuLeaveTimer = null;
-      parentMenu.subMenuLeaveFn = null;
-    }
-  },
-
-  isChildrenSelected() {
+  isChildrenSelected = () => {
     const ret = { find: false };
-    loopMenuItemRecusively(this.props.children, this.props.selectedKeys, ret);
+    loopMenuItemRecursively(this.props.children, this.props.selectedKeys, ret);
     return ret.find;
-  },
-  isOpen() {
+  }
+
+  isOpen = () => {
     return this.props.openKeys.indexOf(this.props.eventKey) !== -1;
-  },
+  }
+
+  adjustWidth = () => {
+    /* istanbul ignore if */
+    if (!this.subMenuTitle || !this.menuInstance) {
+      return;
+    }
+    const popupMenu = ReactDOM.findDOMNode(this.menuInstance);
+    if (popupMenu.offsetWidth >= this.subMenuTitle.offsetWidth) {
+      return;
+    }
+
+    /* istanbul ignore next */
+    popupMenu.style.minWidth = `${this.subMenuTitle.offsetWidth}px`;
+  };
+
+  saveSubMenuTitle = (subMenuTitle) => {
+    this.subMenuTitle = subMenuTitle;
+  }
 
   renderChildren(children) {
     const props = this.props;
     const baseProps = {
       mode: props.mode === 'horizontal' ? 'vertical' : props.mode,
-      visible: this.isOpen(),
+      visible: this.props.isOpen,
       level: props.level + 1,
       inlineIndent: props.inlineIndent,
       focusable: false,
@@ -348,31 +354,70 @@ const SubMenu = createReactClass({
       openTransitionName: props.openTransitionName,
       openAnimation: props.openAnimation,
       onOpenChange: this.onOpenChange,
-      closeSubMenuOnMouseLeave: props.closeSubMenuOnMouseLeave,
-      defaultActiveFirst: this.state.defaultActiveFirst,
+      subMenuOpenDelay: props.subMenuOpenDelay,
+      parentMenu: this,
+      subMenuCloseDelay: props.subMenuCloseDelay,
+      forceSubMenuRender: props.forceSubMenuRender,
+      triggerSubMenuAction: props.triggerSubMenuAction,
+      builtinPlacements: props.builtinPlacements,
+      defaultActiveFirst: props.store.getState()
+        .defaultActiveFirst[getMenuIdFromSubMenuEventKey(props.eventKey)],
       multiple: props.multiple,
       prefixCls: props.rootPrefixCls,
       id: this._menuId,
-      ref: this.saveMenuInstance,
-      count: props.count,
+      manualRef: this.saveMenuInstance,
     };
-    return <SubPopupMenu {...baseProps}>{children}</SubPopupMenu>;
-  },
+
+    const haveRendered = this.haveRendered;
+    this.haveRendered = true;
+
+    this.haveOpened = this.haveOpened || baseProps.visible || baseProps.forceSubMenuRender;
+    // never rendered not planning to, don't render
+    if (!this.haveOpened) {
+      return <div />;
+    }
+
+    // don't show transition on first rendering (no animation for opened menu)
+    // show appear transition if it's not visible (not sure why)
+    // show appear transition if it's not inline mode
+    const transitionAppear = haveRendered || !baseProps.visible || !baseProps.mode === 'inline';
+
+    baseProps.className = ` ${baseProps.prefixCls}-sub`;
+    const animProps = {};
+
+    if (baseProps.openTransitionName) {
+      animProps.transitionName = baseProps.openTransitionName;
+    } else if (typeof baseProps.openAnimation === 'object') {
+      animProps.animation = { ...baseProps.openAnimation };
+      if (!transitionAppear) {
+        delete animProps.animation.appear;
+      }
+    }
+
+    return (
+      <Animate
+        {...animProps}
+        showProp="visible"
+        component=""
+        transitionAppear={transitionAppear}
+      >
+        <SubPopupMenu {...baseProps} id={this._menuId}>{children}</SubPopupMenu>
+      </Animate>
+    );
+  }
 
   render() {
-    const isOpen = this.isOpen();
-    this.haveOpen = this.haveOpen || isOpen;
-    const props = this.props;
+    const props = { ...this.props };
+    const isOpen = props.isOpen;
     const prefixCls = this.getPrefixCls();
-    const classes = {
+    const isInlineMode = props.mode === 'inline';
+    const className = classNames(prefixCls, `${prefixCls}-${props.mode}`, {
       [props.className]: !!props.className,
-      [`${prefixCls}-${props.mode}`]: 1,
-    };
-
-    classes[this.getOpenClassName()] = isOpen;
-    classes[this.getActiveClassName()] = props.active;
-    classes[this.getDisabledClassName()] = props.disabled;
-    classes[this.getSelectedClassName()] = this.isChildrenSelected();
+      [this.getOpenClassName()]: isOpen,
+      [this.getActiveClassName()]: props.active || (isOpen && !isInlineMode),
+      [this.getDisabledClassName()]: props.disabled,
+      [this.getSelectedClassName()]: this.isChildrenSelected(),
+    });
 
     if (!this._menuId) {
       if (props.eventKey) {
@@ -382,8 +427,6 @@ const SubMenu = createReactClass({
       }
     }
 
-    classes[prefixCls] = true;
-    classes[`${prefixCls}-${props.mode}`] = 1;
     let mouseEvents = {};
     let titleClickEvents = {};
     let titleMouseEvents = {};
@@ -392,40 +435,106 @@ const SubMenu = createReactClass({
         onMouseLeave: this.onMouseLeave,
         onMouseEnter: this.onMouseEnter,
       };
+
+      // only works in title, not outer li
       titleClickEvents = {
         onClick: this.onTitleClick,
       };
-      // only works in title, not outer li
       titleMouseEvents = {
         onMouseEnter: this.onTitleMouseEnter,
         onMouseLeave: this.onTitleMouseLeave,
       };
     }
+
     const style = {};
-    if (props.mode === 'inline') {
+    if (isInlineMode) {
       style.paddingLeft = props.inlineIndent * props.level;
     }
+
+    let ariaOwns = {};
+    // only set aria-owns when menu is open
+    // otherwise it would be an invalid aria-owns value
+    // since corresponding node cannot be found
+    if (this.props.isOpen) {
+      ariaOwns = {
+        'aria-owns': this._menuId,
+      };
+    }
+
+    const title = (
+      <div
+        ref={this.saveSubMenuTitle}
+        style={style}
+        className={`${prefixCls}-title`}
+        {...titleMouseEvents}
+        {...titleClickEvents}
+        aria-expanded={isOpen}
+        {...ariaOwns}
+        aria-haspopup="true"
+        title={typeof props.title === 'string' ? props.title : undefined}
+      >
+        {props.title}
+        <i className={`${prefixCls}-arrow`} />
+      </div>
+    );
+    const children = this.renderChildren(props.children);
+
+    const getPopupContainer = props.parentMenu.isRootMenu ?
+      props.parentMenu.props.getPopupContainer : triggerNode => triggerNode.parentNode;
+    const popupPlacement = popupPlacementMap[props.mode];
+    const popupAlign = props.popupOffset ? { offset: props.popupOffset } : {};
+    const popupClassName = props.mode === 'inline' ? '' : props.popupClassName;
+    const {
+      disabled,
+      triggerSubMenuAction,
+      subMenuOpenDelay,
+      forceSubMenuRender,
+      subMenuCloseDelay,
+      builtinPlacements,
+    } = props;
+    menuAllProps.forEach(key => delete props[key]);
+    // Set onClick to null, to ignore propagated onClick event
+    delete props.onClick;
+
     return (
-      <li className={classnames(classes)} {...mouseEvents}>
-        <div
-          style={style}
-          className={`${prefixCls}-title`}
-          {...titleMouseEvents}
-          {...titleClickEvents}
-          aria-expanded={isOpen}
-          aria-owns={this._menuId}
-          aria-haspopup="true"
-          title={typeof props.title === 'string' ? props.title : undefined}
-        >
-          {props.title}
-          <i className={`${prefixCls}-arrow`} />
-        </div>
-        {this.renderChildren(props.children)}
+      <li
+        {...props}
+        {...mouseEvents}
+        className={className}
+        role="menuitem"
+      >
+        {isInlineMode && title}
+        {isInlineMode && children}
+        {!isInlineMode && (
+          <Trigger
+            prefixCls={prefixCls}
+            popupClassName={`${prefixCls}-popup ${popupClassName}`}
+            getPopupContainer={getPopupContainer}
+            builtinPlacements={Object.assign({}, placements, builtinPlacements)}
+            popupPlacement={popupPlacement}
+            popupVisible={isOpen}
+            popupAlign={popupAlign}
+            popup={children}
+            action={disabled ? [] : [triggerSubMenuAction]}
+            mouseEnterDelay={subMenuOpenDelay}
+            mouseLeaveDelay={subMenuCloseDelay}
+            onPopupVisibleChange={this.onPopupVisibleChange}
+            forceRender={forceSubMenuRender}
+          >
+            {title}
+          </Trigger>
+        )}
       </li>
     );
-  },
-});
+  }
+}
 
-SubMenu.isSubMenu = 1;
+const connected = connect(({ openKeys, activeKey, selectedKeys }, { eventKey, subMenuKey }) => ({
+  isOpen: openKeys.indexOf(eventKey) > -1,
+  active: activeKey[subMenuKey] === eventKey,
+  selectedKeys,
+}))(SubMenu);
 
-export default SubMenu;
+connected.isSubMenu = true;
+
+export default connected;