+++ tests/MenuItem.spec.js
@@ -1,8 +1,10 @@
 /* eslint-disable no-undef */
 import React from 'react';
-import { mount } from 'enzyme';
+import { mount, shallow } from 'enzyme';
 import KeyCode from 'rc-util/lib/KeyCode';
-import Menu, { MenuItem } from '../src';
+import Menu, { MenuItem, MenuItemGroup, SubMenu } from '../src';
+
+import { MenuItem as NakedMenuItem } from '../src/MenuItem';
 
 describe('MenuItem', () => {
   describe('disabled', () => {
@@ -34,35 +36,109 @@ describe('MenuItem', () => {
     });
   });
 
-  describe('unmount', () => {
-    const App = ({ show }) => {
-      return (
-        <Menu>
-          {show && (
-            <MenuItem key="1">1</MenuItem>
-          )}
-        </Menu>
-      );
-    };
+  describe('menuItem events', () => {
+    let onMouseEnter;
+    let onMouseLeave;
+    let onItemHover;
+    let wrapper;
+    let instance;
+    const domEvent = { keyCode: 13 };
+    const key = '1';
 
-    it('removes self from selectedKeys', () => {
-      const wrapper = mount(<App show />);
-      wrapper.find('MenuItem').simulate('click');
-      expect(wrapper.find('Menu').node.state.selectedKeys).toEqual(['1']);
-      wrapper.setProps({ show: false });
-      expect(wrapper.find('Menu').node.state.selectedKeys).toEqual([]);
+    beforeEach(() => {
+      onMouseEnter = jest.fn();
+      onMouseLeave = jest.fn();
+      onItemHover = jest.fn();
+
+      wrapper = shallow(
+        <NakedMenuItem
+          eventKey={key}
+          onMouseEnter={onMouseEnter}
+          onItemHover={onItemHover}
+          onMouseLeave={onMouseLeave}
+        >1</NakedMenuItem>);
+      instance = wrapper.instance();
     });
 
-    it('clears mouse leave timer', () => {
-      const wrapper = mount(<App show />);
-      const menu = wrapper.find('Menu');
-      wrapper.find('MenuItem').simulate('mouseEnter');
-      wrapper.find('MenuItem').simulate('mouseLeave');
-      expect(menu.node.menuItemMouseLeaveFn).toBeTruthy();
-      expect(menu.node.menuItemMouseLeaveTimer).toBeTruthy();
-      wrapper.setProps({ show: false });
-      expect(menu.node.menuItemMouseLeaveFn).toBe(null);
-      expect(menu.node.menuItemMouseLeaveTimer).toBe(null);
+    it('on enter key down should trigger mouse click', () => {
+      instance.onClick = jest.fn();
+      instance.onKeyDown(domEvent);
+
+      expect(instance.onClick).toHaveBeenCalledWith(domEvent);
+    });
+
+    it('on mouse enter should trigger props.onItemHover props.onMouseEnter', () => {
+      instance.onMouseEnter(domEvent);
+
+      expect(onItemHover).toHaveBeenCalledWith({ key, hover: true });
+      expect(onMouseEnter).toHaveBeenCalledWith({ key, domEvent });
+    });
+
+    it('on mouse leave should trigger props.onItemHover props.onMouseLeave', () => {
+      instance.onMouseLeave(domEvent);
+
+      expect(onItemHover).toHaveBeenCalledWith({ key, hover: false });
+      expect(onMouseLeave).toHaveBeenCalledWith({ key, domEvent });
+    });
+  });
+
+  describe('rest props', () => {
+    it('onClick event should only trigger 1 time along the component hierarchy', () => {
+      const onClick = jest.fn();
+      const restProps = {
+        onClick,
+        'data-whatever': 'whatever',
+        title: 'title',
+        className: 'className',
+        style: { fontSize: 20 },
+      };
+      const wrapper = mount(
+        <Menu mode="inline" activeKey="1">
+          <MenuItem key="1" {...restProps}>1</MenuItem>
+          <SubMenu {...restProps}>
+            <MenuItem key="2" {...restProps}>3</MenuItem>
+          </SubMenu>
+          <MenuItemGroup {...restProps}>
+            <MenuItem key="3" {...restProps}>4</MenuItem>
+          </MenuItemGroup>
+        </Menu>
+      );
+
+      expect(wrapper.render()).toMatchSnapshot();
+      wrapper.find('MenuItem').at(0).simulate('click');
+      expect(onClick).toHaveBeenCalledTimes(1);
+
+      wrapper.find('SubMenu').at(0).simulate('click');
+      expect(onClick).toHaveBeenCalledTimes(1);
+
+      wrapper.find('MenuItemGroup').at(0).simulate('click');
+      expect(onClick).toHaveBeenCalledTimes(1);
+    });
+  });
+
+  describe('overwrite default role', () => {
+    it('should set role to none if null', () => {
+      const wrapper = shallow(<NakedMenuItem role={null}>test</NakedMenuItem>);
+
+      expect(wrapper.render()).toMatchSnapshot();
+    });
+
+    it('should set role to none if none', () => {
+      const wrapper = shallow(<NakedMenuItem role="none">test</NakedMenuItem>);
+
+      expect(wrapper.render()).toMatchSnapshot();
+    });
+
+    it('should set role to listitem', () => {
+      const wrapper = shallow(<NakedMenuItem role="listitem">test</NakedMenuItem>);
+
+      expect(wrapper.render()).toMatchSnapshot();
+    });
+
+    it('should set role to option', () => {
+      const wrapper = shallow(<NakedMenuItem role="option">test</NakedMenuItem>);
+
+      expect(wrapper.render()).toMatchSnapshot();
     });
   });
 });