+++ HISTORY.md
@@ -1,5 +1,17 @@
 # History
 ----
+## 7.0.2 / 2018-05-03
+
+- Better aria-* attributes support.
+- Improve Menu performance by avoiding unnecessary updates.
+- Minor key shortcut bug fixes.
+- Support passing props through to MenuItem list item
+
+## 6.0.0 / 2017-10-30
+
+- Replace openSubMenuOnMouseEnter/closeSubMenuOnMouseLeave with triggerSubMenuAction
+- Add subMenuCloseDelay and subMenuOpenDelay
+- Refactor with rc-trigger
 
 ## 5.1.0 / 2016-09-12